README.MD
Welcome to Live well, Eat well, be Active with Diabetes (LEAD). The objective of the LEAD app is to assist the underserved, ethnically diverse, and lower socioeconomic population who have diabetes. The LEAD mobile app will assist with self-care management for individuals with diabetes. The LEAD app works in tandem with the LEAD curriculum.